Tablas saldo, JugadorXApuesta. Carga todos los usuarios

+++ app/models/saldo.js
@@ -3,7 +3,7 @@ module.exports = function(sequelize, DataTypes) {
   	'Saldos',
     {
       saldo : {
-        type: DataTypes.INTEGER  
+        type: DataTypes.REAL  
       }
     }
   );
